pyproject.toml
[tool.poetry]
name = "trialtask"
version = "0.1.0"
description = ""
authors = ["Nadiabona <Nadinkotikoff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
envparse = "^0.2.0"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
